+++ src/actions/authActions.js
@@ -1,4 +1,4 @@
-import { AUTH_ERROR, AUTH_USER, UNAUTH_USER } from './types';
+import { AUTH_ERROR, AUTH_USER, UNAUTH_USER, PASSWORD_FAIL, PASSWORD_SUCCESS } from './types';
 import { replace } from 'connected-react-router';
 
 import Config from '../config';
@@ -66,4 +66,30 @@ const tokenGetClaims = (token) => {
     return {};
   }
   return JSON.parse(window.atob(tokenArray[1].replace('-', '+').replace('_', '/')));
+};
+
+export const passwordReset = email => async dispatch => {
+  try {
+    const response = await fetch(`${Config.API_URL + Config.routes.auth.resetPassword}/${email}`, {
+      method: 'GET',
+      headers: {
+        Accept: 'application/json',
+        'Content-Type': 'application/json',
+      },
+    });
+    const data = await response.json();
+    if (!data) throw new Error('Empty response from server');
+    if (data.error) throw new Error(data.error.message);
+    notify('Success! Check your email shortly', "Email has been sent to " + email);
+    dispatch({
+      type: PASSWORD_SUCCESS,
+    });
+  } catch (error) {
+    notify('Error with email!', error.message);
+    dispatch({
+      type: PASSWORD_FAIL,
+      payload: error.message,
+      error: true,
+    });
+  }
 };
